0.3.1, number => double

+++ index.js
@@ -101,8 +101,7 @@ class Parquet {
   constructor() {
     this.typemap = {
       "integer": "int64",
-      "number": "float",
-
+      "number": "double",
       "boolean": "boolean"
     };
   }
